+++ dtu/migrations/meta_migrations/2026-06-24-096631_app_id/down.sql
@@ -1,0 +1,3 @@
+DELETE FROM key_values WHERE key IN ('app_id');
+INSERT INTO key_values (id, key, value) VALUES(1, 'app_id', 'c.arve');
+INSERT INTO key_values (id, key, value) VALUES(2, 'app_pkg', 'c.arve');