+++ dtu/migrations/meta_migrations/2026-06-24-096631_app_id/up.sql
@@ -1,0 +1,2 @@
+DELETE FROM key_values WHERE key IN ('app_id', 'app_pkg');
+INSERT INTO key_values (id, key, value) VALUES(1, 'app_id', 'd.tu');