+++ dtu/src/app/files/app/setup/LoggingBinder.kt
@@ -1,4 +1,4 @@
-package c.arve
+package dtu.lib
 
 import android.os.Binder
 import android.os.Parcel